Refactor archive operations out of local operations, documentation pass
