+++ src/adapters/local_operations.rs
@@ -3,26 +3,21 @@
 #[cfg(test)]
 mod tests;
 
+mod archive;
+
 use std::{
-    cell::{Cell, RefCell},
+    cell::Cell,
     collections::{HashMap, HashSet},
     ffi::{OsStr, OsString},
     future::Future,
     io,
     os::{
         fd::{AsFd, AsRawFd, OwnedFd},
-        unix::{
-            ffi::{OsStrExt, OsStringExt},
-            fs::PermissionsExt,
-        },
+        unix::ffi::{OsStrExt, OsStringExt},
     },
-    path::{Component, Path, PathBuf},
+    path::{Path, PathBuf},
     pin::Pin,
     rc::Rc,
-    sync::{
-        Arc,
-        atomic::{AtomicBool, AtomicUsize, Ordering},
-    },
 };
 
 use gtk::{gio, glib, prelude::*};
@@ -31,10 +26,10 @@ use crate::{
     adapters::{gio_file_for_location, location_for_file},
     model::{FileEntry, Location},
     services::{
-        ArchiveFormat, CancelledOperation, CompressRequest, CreateDirectoryRequest,
-        CreateFileRequest, DeleteRequest, ExtractRequest, LoadHandle, OperationEvent,
-        OperationProvider, OperationRequestId, PasteRequest, RenameRequest, RestoreRequest,
-        RestoreSource, TransferConflict, UndoCopyRequest, UndoMoveRequest, validate_basename,
+        CancelledOperation, CompressRequest, CreateDirectoryRequest, CreateFileRequest,
+        DeleteRequest, ExtractRequest, LoadHandle, OperationEvent, OperationProvider,
+        OperationRequestId, PasteRequest, RenameRequest, RestoreRequest, RestoreSource,
+        TransferConflict, UndoCopyRequest, UndoMoveRequest, validate_basename,
     },
 };
 
@@ -1764,66 +1759,6 @@ fn operation_error_summary(errors: &[String], action: &str) -> String {
     summary
 }
 
-async fn write_staged_archive<F>(
-    destination: &Path,
-    archive_path: &Path,
-    conflict: TransferConflict,
-    cancelled: &AtomicBool,
-    write_archive: F,
-) -> Result<(), ArchiveError>
-where
-    F: FnOnce(std::fs::File) -> Result<(), ArchiveError> + Send + 'static,
-{
-    let published_permissions = if conflict == TransferConflict::ReplaceExisting {
-        match std::fs::symlink_metadata(archive_path) {
-            Ok(metadata) if metadata.file_type().is_file() => Some(metadata.permissions()),
-            Ok(_) => None,
-            Err(error) if error.kind() == io::ErrorKind::NotFound => None,
-            Err(error) => return Err(archive_failed(error)),
-        }
-    } else {
-        None
-    }
-    .unwrap_or_else(umask_adjusted_file_permissions);
-    let mut builder = tempfile::Builder::new();
-    builder
-        .prefix(".strata-compression-")
-        .permissions(std::fs::Permissions::from_mode(0o600));
-    let staged = builder.tempfile_in(destination).map_err(archive_failed)?;
-    let file = staged.reopen().map_err(archive_failed)?;
-    gio::spawn_blocking(move || write_archive(file))
-        .await
-        .map_err(|_| archive_failed("Compression task panicked"))??;
-    check_archive_cancelled(cancelled)?;
-    staged
-        .as_file()
-        .set_permissions(published_permissions)
-        .map_err(archive_failed)?;
-    match conflict {
-        TransferConflict::FailIfExists => staged.persist_noclobber(archive_path),
-        TransferConflict::ReplaceExisting => staged.persist(archive_path),
-    }
-    .map(|_| ())
-    .map_err(archive_failed)
-}
-
-fn umask_adjusted_file_permissions() -> std::fs::Permissions {
-    std::fs::Permissions::from_mode(0o666 & !process_umask())
-}
-
-fn process_umask() -> u32 {
-    // /proc avoids the process-global umask(2) set-and-restore race in a GUI.
-    std::fs::read_to_string("/proc/self/status")
-        .ok()
-        .and_then(|status| {
-            status.lines().find_map(|line| {
-                line.strip_prefix("Umask:")
-                    .and_then(|value| u32::from_str_radix(value.trim(), 8).ok())
-            })
-        })
-        .unwrap_or(0o022)
-}
-
 fn deletion_error_summary(errors: &[String]) -> String {
     operation_error_summary(errors, "deleted")
 }
@@ -2898,1161 +2833,10 @@ impl OperationProvider for LocalOperationProvider {
     }
 
     fn compress(&self, request: CompressRequest, emit: Rc<dyn Fn(OperationEvent)>) -> LoadHandle {
-        let cancelled = Arc::new(AtomicBool::new(false));
-        let task_cancelled = cancelled.clone();
-        let work_cancelled = cancelled.clone();
-        let destination = request.destination.clone();
-        let source_locations = request
-            .entries
-            .iter()
-            .map(|entry| entry.location.clone())
-            .collect::<Vec<_>>();
-        let _task = glib::MainContext::default().spawn_local(async move {
-            let Some(dest_dir) = request.destination.native_path().map(Path::to_path_buf) else {
-                emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: "Archive destination must be a local path".to_owned(),
-                });
-                return;
-            };
-            if let Err(message) = validate_basename(&request.archive_name) {
-                emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: message.to_owned(),
-                });
-                return;
-            }
-            let archive_name = format!("{}.{}", request.archive_name, request.format.extension());
-            let archive_path = dest_dir.join(&archive_name);
-            let entries: Vec<std::path::PathBuf> = request
-                .entries
-                .iter()
-                .filter_map(|e| e.location.native_path().map(Path::to_path_buf))
-                .collect();
-            if entries.is_empty() {
-                emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: "Nothing to compress".to_owned(),
-                });
-                return;
-            }
-            let total = Arc::new(AtomicUsize::new(0));
-            let progress = Arc::new(AtomicUsize::new(0));
-            emit(OperationEvent::ArchiveStarted {
-                request_id: request.id,
-                total: 0,
-            });
-            let timer_id =
-                archive_progress_timer(request.id, &progress, &total, &task_cancelled, &emit);
-            let format = request.format;
-            let password = request.password.clone();
-            let work_progress = progress.clone();
-            let work_total = total.clone();
-            let result = write_staged_archive(
-                &dest_dir,
-                &archive_path,
-                request.conflict,
-                &task_cancelled,
-                move |file| {
-                    let count = count_archive_files(&entries, &work_cancelled)?;
-                    work_total.store(count, Ordering::Relaxed);
-                    match format {
-                        ArchiveFormat::Zip => compress_zip(
-                            file,
-                            &entries,
-                            password.as_deref(),
-                            &work_progress,
-                            &work_cancelled,
-                        ),
-                        ArchiveFormat::SevenZ => compress_7z(
-                            file,
-                            &entries,
-                            password.as_deref(),
-                            &work_progress,
-                            &work_cancelled,
-                        ),
-                        ArchiveFormat::TarGz => {
-                            compress_tar(file, &entries, true, &work_progress, &work_cancelled)
-                        }
-                        ArchiveFormat::Tar => {
-                            compress_tar(file, &entries, false, &work_progress, &work_cancelled)
-                        }
-                    }
-                },
-            )
-            .await;
-            timer_id.remove();
-            match result {
-                Ok(()) => emit(OperationEvent::Compressed {
-                    request_id: request.id,
-                    archive_name: archive_name.clone(),
-                }),
-                Err(ArchiveError::Cancelled) => emit(cancelled_archive_event(
-                    request.id,
-                    destination,
-                    Vec::new(),
-                    Vec::new(),
-                    source_locations,
-                )),
-                Err(ArchiveError::Failed(error)) => emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: error,
-                }),
-            }
-        });
-        LoadHandle::new(move || {
-            cancelled.store(true, Ordering::Relaxed);
-        })
+        archive::compress(request, emit)
     }
 
     fn extract(&self, request: ExtractRequest, emit: Rc<dyn Fn(OperationEvent)>) -> LoadHandle {
-        let cancelled = Arc::new(AtomicBool::new(false));
-        let task_cancelled = cancelled.clone();
-        let work_cancelled = cancelled.clone();
-        let destination = request.destination.clone();
-        let _task = glib::MainContext::default().spawn_local(async move {
-            let Some(archive_path) = request.entry.location.native_path().map(Path::to_path_buf)
-            else {
-                emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: "Archive must be a local file".to_owned(),
-                });
-                return;
-            };
-            let Some(dest_dir) = request.destination.native_path().map(Path::to_path_buf) else {
-                emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: "Extract destination must be a local path".to_owned(),
-                });
-                return;
-            };
-            let format = ArchiveFormat::from_extension(&request.entry.display_name);
-            let password = request.password.clone();
-            let display_name = request.entry.display_name.clone();
-            let progress = Arc::new(AtomicUsize::new(0));
-            let total = Arc::new(AtomicUsize::new(0));
-            emit(OperationEvent::ArchiveStarted {
-                request_id: request.id,
-                total: 0,
-            });
-            let timer_id =
-                archive_progress_timer(request.id, &progress, &total, &task_cancelled, &emit);
-            let work_progress = progress.clone();
-            let work_total = total.clone();
-            let result = gio::spawn_blocking(move || match format {
-                Some(ArchiveFormat::Zip) => {
-                    let file = std::fs::File::open(&archive_path).map_err(|e| e.to_string())?;
-                    let mut archive = zip::ZipArchive::new(file).map_err(|e| e.to_string())?;
-                    work_total.store(archive.len(), Ordering::Relaxed);
-                    extract_zip_from_archive(
-                        &mut archive,
-                        &dest_dir,
-                        password.as_deref(),
-                        &work_progress,
-                        &work_cancelled,
-                    )
-                }
-                Some(ArchiveFormat::SevenZ) => {
-                    let pw = password
-                        .as_deref()
-                        .map(sevenz_rust2::Password::from)
-                        .unwrap_or_default();
-                    let file = std::fs::File::open(&archive_path).map_err(|e| e.to_string())?;
-                    extract_7z_from_reader(file, &dest_dir, pw, &work_progress, &work_cancelled)
-                }
-                Some(ArchiveFormat::TarGz) => extract_tar(
-                    &archive_path,
-                    &dest_dir,
-                    true,
-                    &work_progress,
-                    &work_cancelled,
-                ),
-                Some(ArchiveFormat::Tar) => extract_tar(
-                    &archive_path,
-                    &dest_dir,
-                    false,
-                    &work_progress,
-                    &work_cancelled,
-                ),
-                None => Err(archive_failed(format!(
-                    "Unsupported archive format: {display_name}"
-                ))),
-            })
-            .await;
-            timer_id.remove();
-            match result {
-                Ok(Ok(ArchiveOutcome::Completed(first_name))) => emit(OperationEvent::Extracted {
-                    request_id: request.id,
-                    first_name,
-                }),
-                Ok(Ok(ArchiveOutcome::Cancelled {
-                    completed,
-                    failed,
-                    not_attempted,
-                })) => emit(cancelled_archive_event(
-                    request.id,
-                    destination,
-                    completed,
-                    failed,
-                    not_attempted,
-                )),
-                Ok(Err(ArchiveError::Cancelled)) => emit(cancelled_archive_event(
-                    request.id,
-                    destination,
-                    Vec::new(),
-                    Vec::new(),
-                    Vec::new(),
-                )),
-                Ok(Err(ArchiveError::Failed(error))) => emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: error,
-                }),
-                Err(_) => emit(OperationEvent::Failed {
-                    request_id: request.id,
-                    message: "Extraction task panicked".to_owned(),
-                }),
-            }
-        });
-        LoadHandle::new(move || {
-            cancelled.store(true, Ordering::Relaxed);
-        })
-    }
-}
-
-enum ArchiveSource {
-    File(std::fs::File),
-    Directory(std::fs::File),
-    Symlink(PathBuf),
-}
-
-fn open_archive_source<Fd: AsFd>(parent: &Fd, name: &OsStr) -> Result<ArchiveSource, String> {
-    let stat = rustix::fs::statat(parent, name, rustix::fs::AtFlags::SYMLINK_NOFOLLOW)
-        .map_err(|error| error.to_string())?;
-    match rustix::fs::FileType::from_raw_mode(stat.st_mode) {
-        rustix::fs::FileType::Symlink => {
-            let target = rustix::fs::readlinkat(parent, name, Vec::new())
-                .map_err(|error| error.to_string())?;
-            Ok(ArchiveSource::Symlink(PathBuf::from(OsString::from_vec(
-                target.into_bytes(),
-            ))))
-        }
-        rustix::fs::FileType::Directory => open_local_child_directory(parent, name)
-            .map(std::fs::File::from)
-            .map(ArchiveSource::Directory),
-        rustix::fs::FileType::RegularFile => {
-            let file = rustix::fs::openat2(
-                parent,
-                name,
-                rustix::fs::OFlags::RDONLY
-                    | rustix::fs::OFlags::NONBLOCK
-                    | rustix::fs::OFlags::CLOEXEC,
-                rustix::fs::Mode::empty(),
-                rustix::fs::ResolveFlags::BENEATH
-                    | rustix::fs::ResolveFlags::NO_SYMLINKS
-                    | rustix::fs::ResolveFlags::NO_MAGICLINKS,
-            )
-            .map(std::fs::File::from)
-            .map_err(|error| error.to_string())?;
-            if !file
-                .metadata()
-                .map_err(|error| error.to_string())?
-                .is_file()
-            {
-                return Err("The file type changed during compression".to_owned());
-            }
-            Ok(ArchiveSource::File(file))
-        }
-        _ => Err("Compression supports only regular files, folders, and symbolic links".to_owned()),
-    }
-}
-
-fn visit_archive_entries(
-    entries: &[PathBuf],
-    cancelled: &AtomicBool,
-    visit: &mut impl FnMut(&Path, &ArchiveSource) -> Result<(), ArchiveError>,
-) -> Result<(), ArchiveError> {
-    for entry in entries {
-        check_archive_cancelled(cancelled)?;
-        let name = entry.file_name().ok_or("Entry has no file name")?;
-        let parent = open_local_parent_directory(entry.parent().ok_or("Entry has no parent")?)?;
-        visit_archive_entry(&parent, name, Path::new(name), cancelled, visit)?;
-    }
-    Ok(())
-}
-
-fn visit_archive_entry<Fd: AsFd>(
-    parent: &Fd,
-    name: &OsStr,
-    archive_path: &Path,
-    cancelled: &AtomicBool,
-    visit: &mut impl FnMut(&Path, &ArchiveSource) -> Result<(), ArchiveError>,
-) -> Result<(), ArchiveError> {
-    check_archive_cancelled(cancelled)?;
-    let source = open_archive_source(parent, name).map_err(|error| {
-        archive_failed(format!(
-            "Could not compress {}: {error}",
-            archive_path.display()
-        ))
-    })?;
-    visit(archive_path, &source)?;
-    if let ArchiveSource::Directory(directory) = source {
-        for child in local_directory_children(&directory)? {
-            visit_archive_entry(
-                &directory,
-                &child,
-                &archive_path.join(&child),
-                cancelled,
-                visit,
-            )?;
-        }
-    }
-    Ok(())
-}
-
-fn compress_zip(
-    file: std::fs::File,
-    entries: &[std::path::PathBuf],
-    password: Option<&str>,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<(), ArchiveError> {
-    let writer = std::io::BufWriter::with_capacity(COPY_BUF, file);
-    let mut writer = zip::ZipWriter::new(writer);
-    let deflated = zip::write::SimpleFileOptions::default()
-        .compression_method(zip::CompressionMethod::Deflated)
-        .compression_level(Some(6));
-    let stored =
-        zip::write::SimpleFileOptions::default().compression_method(zip::CompressionMethod::Stored);
-    let deflated = if let Some(pw) = password {
-        deflated.with_aes_encryption(zip::AesMode::Aes256, pw)
-    } else {
-        deflated
-    };
-    let stored = if let Some(pw) = password {
-        stored.with_aes_encryption(zip::AesMode::Aes256, pw)
-    } else {
-        stored
-    };
-    visit_archive_entries(entries, cancelled, &mut |path, source| {
-        let name = path.to_string_lossy();
-        match source {
-            ArchiveSource::Directory(_) => {
-                return writer.add_directory(name, stored).map_err(archive_failed);
-            }
-            ArchiveSource::Symlink(target) => {
-                let target = target.to_str().ok_or_else(|| {
-                    format!(
-                        "ZIP cannot preserve the non-UTF-8 link target of {}. Use TAR instead.",
-                        path.display()
-                    )
-                })?;
-                writer
-                    .add_symlink(name, target, stored)
-                    .map_err(|error| error.to_string())?;
-            }
-            ArchiveSource::File(file) => {
-                let options = if is_incompressible(path) {
-                    stored
-                } else {
-                    deflated
-                };
-                writer
-                    .start_file(name, options)
-                    .map_err(|error| error.to_string())?;
-                copy_with_big_buf(
-                    std::io::BufReader::with_capacity(COPY_BUF, file),
-                    &mut writer,
-                    cancelled,
-                )?;
-            }
-        }
-        progress.fetch_add(1, Ordering::Relaxed);
-        Ok(())
-    })?;
-    check_archive_cancelled(cancelled)?;
-    writer
-        .finish()
-        .map_err(|error| error.to_string())?
-        .into_inner()
-        .map_err(|error| error.to_string())?;
-    Ok(())
-}
-
-fn compress_tar(
-    file: std::fs::File,
-    entries: &[std::path::PathBuf],
-    gzip: bool,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<(), ArchiveError> {
-    let writer = std::io::BufWriter::with_capacity(COPY_BUF, file);
-    if gzip {
-        let mut encoder = flate2::write::GzEncoder::new(writer, flate2::Compression::default());
-        append_tar_entries(&mut encoder, entries, progress, cancelled)?;
-        encoder
-            .finish()
-            .map_err(|error| error.to_string())?
-            .into_inner()
-            .map_err(|error| error.to_string())?;
-    } else {
-        let mut writer = writer;
-        append_tar_entries(&mut writer, entries, progress, cancelled)?;
-        writer.into_inner().map_err(|error| error.to_string())?;
-    }
-    Ok(())
-}
-
-fn append_tar_entries(
-    writer: &mut dyn std::io::Write,
-    entries: &[std::path::PathBuf],
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<(), ArchiveError> {
-    let mut builder = tar::Builder::new(writer);
-    visit_archive_entries(entries, cancelled, &mut |path, source| {
-        let mut header = tar::Header::new_gnu();
-        match source {
-            ArchiveSource::Symlink(target) => {
-                header.set_entry_type(tar::EntryType::Symlink);
-                header.set_size(0);
-                header.set_mode(0o777);
-                header.set_uid(0);
-                header.set_gid(0);
-                builder
-                    .append_link(&mut header, path, target)
-                    .map_err(|error| error.to_string())?;
-            }
-            ArchiveSource::Directory(directory) => {
-                header.set_metadata(&directory.metadata().map_err(|error| error.to_string())?);
-                return builder
-                    .append_data(&mut header, path, std::io::empty())
-                    .map_err(archive_failed);
-            }
-            ArchiveSource::File(file) => {
-                let mut file = file.try_clone().map_err(|error| error.to_string())?;
-                builder
-                    .append_file(path, &mut file)
-                    .map_err(|error| error.to_string())?;
-            }
-        }
-        check_archive_cancelled(cancelled)?;
-        progress.fetch_add(1, Ordering::Relaxed);
-        Ok(())
-    })?;
-    check_archive_cancelled(cancelled)?;
-    builder.finish().map_err(archive_failed)?;
-    Ok(())
-}
-
-fn validated_archive_path(name: &str) -> Result<PathBuf, String> {
-    let normalized = name.replace('\\', "/");
-    if normalized.is_empty() || normalized.starts_with('/') {
-        return Err(format!("Refusing unsafe archive path: {name}"));
-    }
-
-    let mut path = PathBuf::new();
-    for component in normalized.split('/') {
-        match component.as_bytes() {
-            b"" | b"." => {}
-            b".." => return Err(format!("Refusing unsafe archive path: {name}")),
-            bytes if bytes.len() >= 2 && bytes[0].is_ascii_alphabetic() && bytes[1] == b':' => {
-                return Err(format!("Refusing unsafe archive path: {name}"));
-            }
-            _ => path.push(component),
-        }
-    }
-    if path.as_os_str().is_empty() {
-        return Err(format!("Refusing empty archive path: {name}"));
-    }
-    Ok(path)
-}
-
-fn suffixed_name(name: &OsStr, index: u64) -> OsString {
-    let path = Path::new(name);
-    let mut candidate = path.file_stem().unwrap_or(name).as_bytes().to_vec();
-    candidate.extend_from_slice(format!(" ({index})").as_bytes());
-    if let Some(extension) = path.extension() {
-        candidate.push(b'.');
-        candidate.extend_from_slice(extension.as_bytes());
-    }
-    OsString::from_vec(candidate)
-}
-
-struct ExtractionDestination {
-    root: OwnedFd,
-}
-
-impl ExtractionDestination {
-    fn open(path: &Path) -> Result<Self, String> {
-        let root = rustix::fs::open(
-            path,
-            rustix::fs::OFlags::RDONLY
-                | rustix::fs::OFlags::DIRECTORY
-                | rustix::fs::OFlags::NOFOLLOW
-                | rustix::fs::OFlags::CLOEXEC,
-            rustix::fs::Mode::empty(),
-        )
-        .map_err(|error| format!("Could not open extraction destination: {error}"))?;
-        Ok(Self { root })
-    }
-
-    fn available_name<Fd: AsFd>(&self, directory: &Fd, name: &OsStr) -> Result<OsString, String> {
-        for index in 1.. {
-            let candidate = if index == 1 {
-                name.to_owned()
-            } else {
-                suffixed_name(name, index)
-            };
-            match rustix::fs::statat(directory, &candidate, rustix::fs::AtFlags::SYMLINK_NOFOLLOW) {
-                Err(rustix::io::Errno::NOENT) => return Ok(candidate),
-                Err(error) => {
-                    return Err(format!(
-                        "Could not inspect extraction path {}: {error}",
-                        candidate.to_string_lossy()
-                    ));
-                }
-                Ok(stat) => match rustix::fs::FileType::from_raw_mode(stat.st_mode) {
-                    rustix::fs::FileType::RegularFile | rustix::fs::FileType::Directory => {}
-                    _ => {
-                        return Err(format!(
-                            "Refusing to extract over special filesystem object: {}",
-                            candidate.to_string_lossy()
-                        ));
-                    }
-                },
-            }
-        }
-        Err(format!(
-            "Could not find an available extraction name for {}",
-            name.to_string_lossy()
-        ))
-    }
-
-    fn create_directories(&self, path: &Path) -> Result<OwnedFd, String> {
-        let mut directory = self.root.try_clone().map_err(|error| error.to_string())?;
-        for component in path.components() {
-            let Component::Normal(name) = component else {
-                return Err("Invalid internal extraction path".to_owned());
-            };
-            match rustix::fs::mkdirat(&directory, name, rustix::fs::Mode::from_raw_mode(0o777)) {
-                Ok(()) | Err(rustix::io::Errno::EXIST) => {}
-                Err(error) => return Err(error.to_string()),
-            }
-            directory = rustix::fs::openat(
-                &directory,
-                name,
-                rustix::fs::OFlags::RDONLY
-                    | rustix::fs::OFlags::DIRECTORY
-                    | rustix::fs::OFlags::NOFOLLOW
-                    | rustix::fs::OFlags::CLOEXEC,
-                rustix::fs::Mode::empty(),
-            )
-            .map_err(|error| error.to_string())?;
-        }
-        Ok(directory)
-    }
-
-    fn create_file(&self, path: &Path) -> Result<(std::fs::File, PathBuf), String> {
-        let parent = self.create_directories(path.parent().unwrap_or_else(|| Path::new("")))?;
-        let name = path
-            .file_name()
-            .ok_or_else(|| "Archive entry has no file name".to_owned())?;
-        let name = self.available_name(&parent, name)?;
-        let mut created = PathBuf::new();
-        if let Some(parent_path) = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            created.push(parent_path);
-        }
-        created.push(&name);
-        let file = rustix::fs::openat(
-            parent,
-            name,
-            rustix::fs::OFlags::WRONLY
-                | rustix::fs::OFlags::CREATE
-                | rustix::fs::OFlags::EXCL
-                | rustix::fs::OFlags::NOFOLLOW
-                | rustix::fs::OFlags::CLOEXEC,
-            rustix::fs::Mode::from_raw_mode(0o666),
-        )
-        .map(std::fs::File::from)
-        .map_err(|error| error.to_string())?;
-        Ok((file, created))
-    }
-
-    fn remove_file(&self, path: &Path) -> Result<(), String> {
-        let parent = self.create_directories(path.parent().unwrap_or_else(|| Path::new("")))?;
-        let name = path
-            .file_name()
-            .ok_or_else(|| "Archive entry has no file name".to_owned())?;
-        rustix::fs::unlinkat(&parent, name, rustix::fs::AtFlags::empty()).map_err(|error| {
-            format!(
-                "Could not remove incomplete extraction {}: {error}",
-                path.display()
-            )
-        })
-    }
-}
-
-fn count_archive_files(entries: &[PathBuf], cancelled: &AtomicBool) -> Result<usize, ArchiveError> {
-    let mut count = 0;
-    visit_archive_entries(entries, cancelled, &mut |_, source| {
-        if !matches!(source, ArchiveSource::Directory(_)) {
-            count += 1;
-        }
-        Ok(())
-    })?;
-    Ok(count)
-}
-
-const COPY_BUF: usize = 1 << 20; // 1 MiB
-const ARCHIVE_CANCELLED: &str = "Operation cancelled";
-
-#[derive(Debug, PartialEq, Eq)]
-enum ArchiveError {
-    Cancelled,
-    Failed(String),
-}
-
-impl std::fmt::Display for ArchiveError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Cancelled => f.write_str(ARCHIVE_CANCELLED),
-            Self::Failed(message) => f.write_str(message),
-        }
-    }
-}
-
-impl std::error::Error for ArchiveError {}
-
-impl From<String> for ArchiveError {
-    fn from(message: String) -> Self {
-        Self::Failed(message)
-    }
-}
-
-impl From<&str> for ArchiveError {
-    fn from(message: &str) -> Self {
-        Self::Failed(message.to_owned())
-    }
-}
-
-fn archive_failed(error: impl std::fmt::Display) -> ArchiveError {
-    ArchiveError::Failed(error.to_string())
-}
-
-fn sevenz_cancelled() -> sevenz_rust2::Error {
-    sevenz_rust2::Error::Other(ARCHIVE_CANCELLED.into())
-}
-
-fn sevenz_is_cancelled(error: &sevenz_rust2::Error) -> bool {
-    matches!(error, sevenz_rust2::Error::Other(message) if message.as_ref() == ARCHIVE_CANCELLED)
-}
-
-enum ArchiveOutcome<T> {
-    Completed(T),
-    Cancelled {
-        completed: Vec<Location>,
-        failed: Vec<Location>,
-        not_attempted: Vec<Location>,
-    },
-}
-
-fn check_archive_cancelled(cancelled: &AtomicBool) -> Result<(), ArchiveError> {
-    if cancelled.load(Ordering::Relaxed) {
-        Err(ArchiveError::Cancelled)
-    } else {
-        Ok(())
-    }
-}
-
-fn cancelled_archive_event(
-    request_id: OperationRequestId,
-    destination: Location,
-    completed: Vec<Location>,
-    failed: Vec<Location>,
-    not_attempted: Vec<Location>,
-) -> OperationEvent {
-    OperationEvent::Cancelled {
-        request_id,
-        result: CancelledOperation {
-            completed,
-            failed,
-            not_attempted,
-            affected_locations: HashSet::from([destination]),
-        },
-    }
-}
-
-fn extract_entry_location(destination: &Path, relative: &Path) -> Location {
-    Location::local(destination.join(relative))
-}
-
-fn zip_entry_locations(
-    archive: &zip::ZipArchive<std::fs::File>,
-    destination: &Path,
-    from: usize,
-) -> Vec<Location> {
-    (from..archive.len())
-        .filter_map(|index| archive.name_for_index(index))
-        .map(|name| Location::local(destination.join(name)))
-        .collect()
-}
-
-fn tar_entry_location<'a, R: std::io::Read + 'a>(
-    entry: tar::Entry<'a, R>,
-    dest_dir: &Path,
-) -> Option<Location> {
-    let name = entry.path().ok()?;
-    let path = validated_archive_path(&name.to_string_lossy()).ok()?;
-    Some(extract_entry_location(dest_dir, &path))
-}
-
-fn sevenz_locations_from(
-    dest_dir: &Path,
-    names: &[String],
-    from_name: &str,
-    skip_current: bool,
-) -> Vec<Location> {
-    let Some(index) = names.iter().position(|name| name == from_name) else {
-        return Vec::new();
-    };
-    let start = if skip_current { index + 1 } else { index };
-    names
-        .get(start..)
-        .unwrap_or(&[])
-        .iter()
-        .filter_map(|name| validated_archive_path(name).ok())
-        .map(|path| extract_entry_location(dest_dir, &path))
-        .collect()
-}
-
-fn cancelled_extract_after_partial_write(
-    dest_dir: &Path,
-    created: &Path,
-    completed: Vec<Location>,
-    remaining: Vec<Location>,
-    removed: Result<(), String>,
-) -> ArchiveOutcome<Option<String>> {
-    let interrupted = extract_entry_location(dest_dir, created);
-    if removed.is_ok() {
-        let mut not_attempted = vec![interrupted];
-        not_attempted.extend(remaining);
-        ArchiveOutcome::Cancelled {
-            completed,
-            failed: Vec::new(),
-            not_attempted,
-        }
-    } else {
-        ArchiveOutcome::Cancelled {
-            completed,
-            failed: vec![interrupted],
-            not_attempted: remaining,
-        }
-    }
-}
-
-fn copy_with_big_buf(
-    mut reader: impl std::io::Read,
-    writer: &mut (impl std::io::Write + ?Sized),
-    cancelled: &AtomicBool,
-) -> Result<u64, ArchiveError> {
-    let mut buf = vec![0u8; COPY_BUF];
-    let mut total = 0;
-    loop {
-        check_archive_cancelled(cancelled)?;
-        let n = reader.read(&mut buf).map_err(archive_failed)?;
-        if n == 0 {
-            break;
-        }
-        writer.write_all(&buf[..n]).map_err(archive_failed)?;
-        total += n as u64;
-    }
-    Ok(total)
-}
-
-/// Spawns a 100ms timer that polls progress counters and emits ArchiveProgress events.
-fn archive_progress_timer(
-    request_id: OperationRequestId,
-    progress: &Arc<AtomicUsize>,
-    total: &Arc<AtomicUsize>,
-    cancelled: &Arc<AtomicBool>,
-    emit: &Rc<dyn Fn(OperationEvent)>,
-) -> glib::SourceId {
-    let timer_progress = progress.clone();
-    let timer_total = total.clone();
-    let timer_cancelled = cancelled.clone();
-    let timer_emit = emit.clone();
-    glib::timeout_add_local(std::time::Duration::from_millis(100), move || {
-        // Keep the source until the task calls remove(); Break would double-remove.
-        if !timer_cancelled.load(Ordering::Relaxed) {
-            timer_emit(OperationEvent::ArchiveProgress {
-                request_id,
-                completed: timer_progress.load(Ordering::Relaxed),
-                total: timer_total.load(Ordering::Relaxed),
-            });
-        }
-        glib::ControlFlow::Continue
-    })
-}
-
-/// File extensions that are already compressed — storing them raw saves CPU with zero size gain.
-const INCOMPRESSIBLE_EXTS: &[&str] = &[
-    "zip", "7z", "gz", "bz2", "xz", "zst", "tar", "rar", "lz", "lz4", "br", "mp4", "mkv", "avi",
-    "mov", "webm", "flv", "wmv", "jpg", "jpeg", "png", "webp", "gif", "heic", "avif", "bmp", "mp3",
-    "flac", "aac", "ogg", "opus", "wma", "m4a", "pdf", "epub", "docx", "xlsx", "pptx", "odt",
-    "ods", "odp", "iso", "dmg", "deb", "rpm", "apk", "jar", "war",
-];
-
-fn is_incompressible(path: &Path) -> bool {
-    path.extension()
-        .and_then(|e| e.to_str())
-        .map(|e| INCOMPRESSIBLE_EXTS.contains(&e.to_lowercase().as_str()))
-        .unwrap_or(false)
-}
-
-/// Tracks renamed top-level entries so child paths follow the rename.
-struct ExtractNameResolver {
-    renames: std::collections::HashMap<OsString, OsString>,
-}
-
-impl ExtractNameResolver {
-    fn new() -> Self {
-        Self {
-            renames: std::collections::HashMap::new(),
-        }
-    }
-
-    /// Resolves a validated relative entry path to a conflict-free relative path.
-    /// If the top-level component already exists, it's renamed to "name (2)", etc.
-    fn resolve(
-        &mut self,
-        destination: &ExtractionDestination,
-        path: &Path,
-    ) -> Result<PathBuf, String> {
-        let top = path
-            .components()
-            .next()
-            .and_then(|component| match component {
-                Component::Normal(name) => Some(name),
-                _ => None,
-            })
-            .ok_or_else(|| "Archive entry has no file name".to_owned())?;
-        let resolved_top = if let Some(existing) = self.renames.get(top) {
-            existing.clone()
-        } else {
-            let name = destination.available_name(&destination.root, top)?;
-            self.renames.insert(top.to_owned(), name.clone());
-            name
-        };
-        let mut resolved = PathBuf::from(resolved_top);
-        resolved.extend(
-            path.components()
-                .skip(1)
-                .map(|component| component.as_os_str()),
-        );
-        Ok(resolved)
-    }
-}
-
-fn extract_zip_from_archive(
-    archive: &mut zip::ZipArchive<std::fs::File>,
-    dest_dir: &Path,
-    password: Option<&str>,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<ArchiveOutcome<Option<String>>, ArchiveError> {
-    let destination = ExtractionDestination::open(dest_dir)?;
-    let pw_bytes = password.map(|p| p.as_bytes());
-    let mut resolver = ExtractNameResolver::new();
-    let mut first_name = None;
-    let mut completed = Vec::new();
-    for i in 0..archive.len() {
-        if cancelled.load(Ordering::Relaxed) {
-            return Ok(ArchiveOutcome::Cancelled {
-                completed,
-                failed: Vec::new(),
-                not_attempted: zip_entry_locations(archive, dest_dir, i),
-            });
-        }
-        let read_options = zip::read::ZipReadOptions::new().password(pw_bytes);
-        let mut entry = archive
-            .by_index_with_options(i, read_options)
-            .map_err(archive_failed)?;
-        let name = entry.name();
-        entry
-            .enclosed_name()
-            .ok_or_else(|| format!("Refusing unsafe ZIP path: {name}"))?;
-        let path = validated_archive_path(name)?;
-        let outpath = resolver.resolve(&destination, &path)?;
-        if first_name.is_none() {
-            first_name = outpath
-                .components()
-                .next()
-                .map(|c| c.as_os_str().to_string_lossy().to_string());
-        }
-        if entry.is_dir() {
-            destination.create_directories(&outpath)?;
-        } else {
-            let (mut outfile, created) = destination.create_file(&outpath)?;
-            if let Err(error) = copy_with_big_buf(&mut entry, &mut outfile, cancelled) {
-                drop(outfile);
-                drop(entry);
-                let removed = destination.remove_file(&created);
-                return match error {
-                    ArchiveError::Cancelled => Ok(cancelled_extract_after_partial_write(
-                        dest_dir,
-                        &created,
-                        completed,
-                        zip_entry_locations(archive, dest_dir, i + 1),
-                        removed,
-                    )),
-                    failed => Err(failed),
-                };
-            }
-        }
-        completed.push(extract_entry_location(dest_dir, &outpath));
-        progress.fetch_add(1, Ordering::Relaxed);
-    }
-    Ok(ArchiveOutcome::Completed(first_name))
-}
-
-fn extract_tar(
-    archive_path: &Path,
-    dest_dir: &Path,
-    gzip: bool,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<ArchiveOutcome<Option<String>>, ArchiveError> {
-    let destination = ExtractionDestination::open(dest_dir)?;
-    let file = std::fs::File::open(archive_path).map_err(archive_failed)?;
-    let reader: Box<dyn std::io::Read> = if gzip {
-        Box::new(flate2::read::GzDecoder::new(file))
-    } else {
-        Box::new(file)
-    };
-    let mut archive = tar::Archive::new(reader);
-    let mut resolver = ExtractNameResolver::new();
-    let mut first_name = None;
-    let mut completed = Vec::new();
-    let entries = archive.entries().map_err(archive_failed)?;
-    for entry in entries {
-        if cancelled.load(Ordering::Relaxed) {
-            let not_attempted = entry
-                .ok()
-                .and_then(|entry| tar_entry_location(entry, dest_dir))
-                .into_iter()
-                .collect();
-            return Ok(ArchiveOutcome::Cancelled {
-                completed,
-                failed: Vec::new(),
-                not_attempted,
-            });
-        }
-        let mut entry = entry.map_err(archive_failed)?;
-        let name = entry.path().map_err(archive_failed)?;
-        if entry.header().entry_type().is_dir() && name == Path::new(".") {
-            continue;
-        }
-        let path = validated_archive_path(&name.to_string_lossy())?;
-        let outpath = resolver.resolve(&destination, &path)?;
-        if first_name.is_none() {
-            first_name = outpath
-                .components()
-                .next()
-                .map(|c| c.as_os_str().to_string_lossy().to_string());
-        }
-        if entry.header().entry_type().is_dir() {
-            destination.create_directories(&outpath)?;
-        } else {
-            let (mut outfile, created) = destination.create_file(&outpath)?;
-            if let Err(error) = copy_with_big_buf(&mut entry, &mut outfile, cancelled) {
-                drop(outfile);
-                drop(entry);
-                let removed = destination.remove_file(&created);
-                return match error {
-                    ArchiveError::Cancelled => Ok(cancelled_extract_after_partial_write(
-                        dest_dir,
-                        &created,
-                        completed,
-                        Vec::new(),
-                        removed,
-                    )),
-                    failed => Err(failed),
-                };
-            }
-        }
-        completed.push(extract_entry_location(dest_dir, &outpath));
-        progress.fetch_add(1, Ordering::Relaxed);
-    }
-    Ok(ArchiveOutcome::Completed(first_name))
-}
-
-fn compress_7z(
-    file: std::fs::File,
-    entries: &[std::path::PathBuf],
-    password: Option<&str>,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<(), ArchiveError> {
-    use sevenz_rust2::encoder_options::{AesEncoderOptions, EncoderOptions, Lzma2Options};
-    let mut writer = sevenz_rust2::ArchiveWriter::new(file).map_err(|e| e.to_string())?;
-    let threads = std::thread::available_parallelism()
-        .map(|n| n.get() as u32)
-        .unwrap_or(1);
-    let lzma2 =
-        sevenz_rust2::EncoderConfiguration::new(sevenz_rust2::EncoderMethod::LZMA2).with_options(
-            EncoderOptions::Lzma2(Lzma2Options::from_level_mt(6, threads, 1 << 26)),
-        );
-    if let Some(pw) = password {
-        let methods = vec![lzma2, AesEncoderOptions::new(pw.into()).into()];
-        writer.set_content_methods(methods);
-    } else {
-        writer.set_content_methods(vec![lzma2]);
-    }
-    visit_archive_entries(entries, cancelled, &mut |path, source| {
-        let name = path.to_string_lossy();
-        let (mut entry, file) = match source {
-            ArchiveSource::Symlink(_) => {
-                return Err(archive_failed(format!(
-                    "7z compression does not support symbolic links: {}. Use ZIP or TAR instead.",
-                    path.display()
-                )));
-            }
-            ArchiveSource::Directory(file) => {
-                (sevenz_rust2::ArchiveEntry::new_directory(&name), file)
-            }
-            ArchiveSource::File(file) => (sevenz_rust2::ArchiveEntry::new_file(&name), file),
-        };
-        let metadata = file.metadata().map_err(|error| error.to_string())?;
-        if let Ok(modified) = metadata.modified()
-            && let Ok(date) = sevenz_rust2::NtTime::try_from(modified)
-        {
-            entry.last_modified_date = date;
-            entry.has_last_modified_date = u64::from(date) > 0;
-        }
-        if let Ok(created) = metadata.created()
-            && let Ok(date) = sevenz_rust2::NtTime::try_from(created)
-        {
-            entry.creation_date = date;
-            entry.has_creation_date = u64::from(date) > 0;
-        }
-        if let Ok(accessed) = metadata.accessed()
-            && let Ok(date) = sevenz_rust2::NtTime::try_from(accessed)
-        {
-            entry.access_date = date;
-            entry.has_access_date = u64::from(date) > 0;
-        }
-        let reader = if matches!(source, ArchiveSource::Directory(_)) {
-            None
-        } else {
-            Some(file)
-        };
-        writer
-            .push_archive_entry(entry, reader)
-            .map_err(archive_failed)?;
-        check_archive_cancelled(cancelled)?;
-        if reader.is_some() {
-            progress.fetch_add(1, Ordering::Relaxed);
-        }
-        Ok(())
-    })?;
-    check_archive_cancelled(cancelled)?;
-    writer.finish().map_err(archive_failed)?;
-    Ok(())
-}
-
-fn extract_7z_from_reader(
-    reader: std::fs::File,
-    dest_dir: &Path,
-    password: sevenz_rust2::Password,
-    progress: &Arc<AtomicUsize>,
-    cancelled: &AtomicBool,
-) -> Result<ArchiveOutcome<Option<String>>, ArchiveError> {
-    let destination = ExtractionDestination::open(dest_dir)?;
-    let mut archive = sevenz_rust2::ArchiveReader::new(reader, password).map_err(archive_failed)?;
-    let entry_names: Vec<String> = archive
-        .archive()
-        .files
-        .iter()
-        .map(|entry| entry.name.clone())
-        .collect();
-    let resolver = RefCell::new(ExtractNameResolver::new());
-    let first_name = RefCell::new(None::<String>);
-    let completed = RefCell::new(Vec::new());
-    let failed = RefCell::new(Vec::new());
-    let not_attempted = RefCell::new(Vec::new());
-    let progress = progress.clone();
-    let dest_dir = dest_dir.to_path_buf();
-    let result = archive.for_each_entries(|entry, reader| {
-        if cancelled.load(Ordering::Relaxed) {
-            *not_attempted.borrow_mut() =
-                sevenz_locations_from(&dest_dir, &entry_names, &entry.name, false);
-            return Err(sevenz_cancelled());
-        }
-        let path = validated_archive_path(&entry.name)
-            .map_err(|error| sevenz_rust2::Error::Other(error.into()))?;
-        let outpath = resolver
-            .borrow_mut()
-            .resolve(&destination, &path)
-            .map_err(|error| sevenz_rust2::Error::Other(error.into()))?;
-        if first_name.borrow().is_none() {
-            *first_name.borrow_mut() = outpath
-                .components()
-                .next()
-                .map(|c| c.as_os_str().to_string_lossy().to_string());
-        }
-        if entry.is_directory {
-            destination
-                .create_directories(&outpath)
-                .map_err(|error| sevenz_rust2::Error::Other(error.into()))?;
-        } else {
-            let (mut file, created) = destination
-                .create_file(&outpath)
-                .map_err(|error| sevenz_rust2::Error::Other(error.into()))?;
-            if let Err(error) = copy_with_big_buf(reader, &mut file, cancelled) {
-                drop(file);
-                let removed = destination.remove_file(&created);
-                return match error {
-                    ArchiveError::Cancelled => {
-                        if removed.is_err() {
-                            failed
-                                .borrow_mut()
-                                .push(extract_entry_location(&dest_dir, &created));
-                            *not_attempted.borrow_mut() =
-                                sevenz_locations_from(&dest_dir, &entry_names, &entry.name, true);
-                        } else {
-                            let mut remaining = vec![extract_entry_location(&dest_dir, &created)];
-                            remaining.extend(sevenz_locations_from(
-                                &dest_dir,
-                                &entry_names,
-                                &entry.name,
-                                true,
-                            ));
-                            *not_attempted.borrow_mut() = remaining;
-                        }
-                        Err(sevenz_cancelled())
-                    }
-                    ArchiveError::Failed(message) => {
-                        Err(sevenz_rust2::Error::Other(message.into()))
-                    }
-                };
-            }
-        }
-        completed
-            .borrow_mut()
-            .push(extract_entry_location(&dest_dir, &outpath));
-        progress.fetch_add(1, Ordering::Relaxed);
-        Ok(true)
-    });
-    match result {
-        Ok(()) => Ok(ArchiveOutcome::Completed(first_name.into_inner())),
-        Err(error) if sevenz_is_cancelled(&error) => Ok(ArchiveOutcome::Cancelled {
-            completed: completed.into_inner(),
-            failed: failed.into_inner(),
-            not_attempted: not_attempted.into_inner(),
-        }),
-        Err(error) => Err(archive_failed(error)),
+        archive::extract(request, emit)
     }
 }
